newscrapper/requirements.txt
amqp==5.0.6
asgiref==3.4.1
beautifulsoup4==4.9.3
billiard==3.6.4.0
boto3==1.18.25
botocore==1.21.25
bs4==0.0.1
celery==5.1.2
certifi==2021.5.30
charset-normalizer==2.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
Django==3.2.6
django-celery-beat==2.2.1
django-cors-headers==3.8.0
django-timezone-field==4.2.1
djangorestframework==3.12.4
dnspython==1.16.0
idna==3.2
jmespath==0.10.0
kombu==5.1.0
prompt-toolkit==3.0.20
pymongo==3.12.0
python-crontab==2.5.1
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
s3transfer==0.5.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
urllib3==1.26.6
vine==5.0.0
wcwidth==0.2.5
